Weapons List Axes through Daggers (#2)
* Added WIP Weapons Tab

Added a tab for weapons, as well as filled in the information for daggers, including links. I followed the formatting of the Dark Souls 3 guide as a reference.

* Axes and Ballistas Added

Added all Axes and Ballistas to data.yaml, and begun adding them to the index

* Fixed Formatting + Links for Weapons

Fixed my formatting for the weapons. Should be fully complete for Axes, Ballistas, and Daggers.

* Small fixup and ran generate.py

* Update data.yaml

* Update data.yaml

* Update data.yaml

* Small fix + ran generate.py

Co-authored-by: Ben Lambeth <accounts@lambethben.com>